Fixed the error WP_Plugins_List_Table occurs when the plugin is activated
The error WP_Plugins_List_Table shows from WP 6.5 to 6.5.3 due to this: https://core.trac.wordpress.org/ticket/61331. However, this error won't occur from WP version 6.5.4 as the WP team reverted it. So, adding a condition to check whether the class exists to redirect to the admin URL if it gets introduced again.